hwcomposer: fix warnings
- remove unused variables and functions
- avoid comparison between different types
- fix old-style field designator extension
- fix wrong print format
- unhide an overloaded virtual function
- add flag -Wno-unused-parameter to disable these ones

Bug: 110425202

+++ merrifield/common/base/Hwcomposer.cpp
@@ -162,7 +162,7 @@ bool Hwcomposer::setPowerMode(int disp, int mode)
         return false;
     }
 
-    if(disp >= mDisplayDevices.size()){
+    if(disp >= (int)mDisplayDevices.size()){
         ETRACE("no device found");
         return false;
     }
@@ -261,7 +261,7 @@ bool Hwcomposer::getDisplayConfigs(int disp,
         return false;
     }
 
-    if(disp >= mDisplayDevices.size()){
+    if(disp >= (int)mDisplayDevices.size()){
         ETRACE("no device found");
         return false;
     }
@@ -286,7 +286,7 @@ bool Hwcomposer::getDisplayAttributes(int disp,
         ETRACE("invalid disp %d", disp);
         return false;
     }
-    if(disp >= mDisplayDevices.size()){
+    if(disp >= (int)mDisplayDevices.size()){
         ETRACE("no device found");
         return false;
     }
@@ -312,7 +312,7 @@ bool Hwcomposer::compositionComplete(int disp)
 
     mDisplayContext->compositionComplete();
 
-    if(disp >= mDisplayDevices.size()){
+    if(disp >= (int)mDisplayDevices.size()){
         ETRACE("no device found");
         return false;
     }
